Added a demo for whisper.
